backend/requirements.txt
flask>=3.0.0
flask_sqlalchemy>=3.1.0
flask_jwt_extended>=4.6.0
flask_migrate>=4.0.5
pymysql>=1.1.0
mysql-connector-python>=8.0.33  # For standalone database setup script
werkzeug>=3.0.1
flask_cors>=4.0.0
flask-wtf>=1.2.1  # Added for CSRF protection
flask-caching>=2.1.0  # Added for performance optimization
Pillow>=10.2.0
numpy>=1.26.0
tensorflow>=2.15.0
email-validator>=2.1.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
bleach>=6.1.0  # Added for HTML sanitization
requests>=2.32.0
urllib3>=1.26.18  # Corrected version

# Machine Learning dependencies for crop prediction
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.1.0

# Security-related
flask-talisman>=1.1.0  # For security headers
flask-limiter>=3.5.0  # For rate limiting
blinker>=1.7.0       # For improved Flask signals and logging

# Optional: for email/notifications
Flask-Mail>=0.9.1

# Optional: for testing and dev
pytest>=8.0.0
